Change typecast to float64
An empty interface for a number defaults to a float64 even though it is not a decimal value.
Therefore you must typecast to float64 and then typecast to an int.

+++ pkg/rapididentity/GetAuthenticationPoliciesForUser.go
@@ -54,7 +54,7 @@ func (ap *AuthenticationPolicy) UnmarshalJSON(b []byte) error {
 		case "id":
 			ap.Id = apv.(string)
 		case "version":
-			ap.Version = apv.(int)
+			ap.Version = int(apv.(float64))
 		case "name":
 			ap.Name = apv.(string)
 		case "enabled":
